Move TypeOfPreview type to interfaces.ts. Fix bug where language comparision of concepts would not appear in master. Removed superfluous ConceptPreviewType and visual element API calls.
